server/requirements.txt
fastapi>=0.100.0,<1.0.0
hypercorn>=0.16.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.0.0,<3.0.0

# LLM Provider dependencies
anthropic>=0.25.0,<1.0.0
openai>=1.0.0,<2.0.0
google-generativeai>=0.3.0,<1.0.0

# File processing
python-docx>=0.8.11,<2.0.0

# Optional dependencies for extended functionality
pyyaml>=6.0,<7.0
datasets>=2.0.0,<4.0.0
numpy>=1.21.0,<3.0.0
hgtk==0.2.1

# Required for file uploads in FastAPI
python-multipart>=0.0.5,<1.0.0

# Retry logic
tenacity>=8.0.0,<9.0.0

# HTTP client for API integrations
httpx>=0.25.0,<1.0.0

# Async file operations
aiofiles>=23.0.0,<25.0.0

# Redis for distributed memory
redis>=5.0.0,<6.0.0

# Observability
structlog>=23.0.0,<25.0.0
prometheus-client>=0.19.0,<1.0.0
notion-client==2.3.0

# GraphQL dependencies
strawberry-graphql[fastapi,subscriptions]>=0.220.0,<1.0.0